refactor: Move the solver tests to the physics directory
