+++ src/serac/physics/CMakeLists.txt
@@ -9,6 +9,10 @@ add_subdirectory(operators)
 add_subdirectory(integrators)
 add_subdirectory(materials)
 
+if(ENABLE_TESTS)
+    add_subdirectory(tests)
+endif()
+
 set(physics_sources
     base_physics.cpp
     solid.cpp
