+++ tests/CMakeLists.txt
@@ -13,25 +13,14 @@ if (ENABLE_GTEST)
     set(test_dependencies serac_physics serac_coefficients test_utils)
     blt_list_append( TO test_dependencies ELEMENTS caliper IF ${SERAC_USE_CALIPER} )
 
-    blt_add_library(
-        NAME        test_utils
-        SOURCES     test_utilities.cpp
-        HEADERS     test_utilities.hpp
-        DEPENDS_ON  serac_physics gtest
-        )
-
-    set(solver_tests
-        serac_solid.cpp
-        serac_solid_reuse_mesh.cpp
-        serac_thermal_solver.cpp
-        serac_thermal_structural_solver.cpp
+    set(solver_utility_tests
         serac_dtor.cpp
         serac_boundary_cond.cpp
         serac_mesh.cpp
         mfem_ex9p_blockilu.cpp
         serac_newmark_test.cpp)
 
-    foreach(filename ${solver_tests})
+    foreach(filename ${solver_utility_tests})
         get_filename_component(test_name ${filename} NAME_WE)
 
         blt_add_executable(NAME        ${test_name}
