+++ tests/serac_component_bc.cpp
@@ -16,7 +16,7 @@
 #include "serac/numerics/mesh_utils.hpp"
 #include "serac/physics/utilities/state_manager.hpp"
 #include "serac/serac_config.hpp"
-#include "test_utilities.hpp"
+#include "serac/physics/tests/test_utilities.hpp"
 
 namespace serac {
